refactor: add loops em controller
